+++ packages/store/src/decorators/selector/selector.ts
@@ -11,29 +11,30 @@ export function Selector<T>(selectors?: T[]): SelectorType<T> {
     key: string | symbol,
     descriptor: TypedPropertyDescriptor<SelectorSpec<T, U>>
   ): TypedPropertyDescriptor<SelectorSpec<T, U>> | void => {
+    descriptor ||= Object.getOwnPropertyDescriptor(target, key)!;
+
+    const originalFn = descriptor?.value;
+
     // Caretaker note: we have still left the `typeof` condition in order to avoid
     // creating a breaking change for projects that still use the View Engine.
     if (typeof ngDevMode === 'undefined' || ngDevMode) {
-      const isNotMethod = !(descriptor && descriptor.value !== null);
-
-      if (isNotMethod) {
+      if (originalFn && typeof originalFn !== 'function') {
         throwSelectorDecoratorError();
       }
     }
 
-    const originalFn = descriptor.value;
     const memoizedFn = createSelector(selectors, originalFn as any, {
       containerClass: target,
       selectorName: key.toString(),
       getSelectorOptions() {
         return {};
-      }
+      },
     });
     const newDescriptor = {
       configurable: true,
       get() {
         return memoizedFn;
-      }
+      },
     };
     // Add hidden property to descriptor
     (<any>newDescriptor)['originalFn'] = originalFn;