+++ packages/store/tests/selector.spec.ts
@@ -20,8 +20,8 @@ describe('Selector', () => {
     name: 'counter',
     defaults: {
       foo: 'Hello',
-      bar: 'World'
-    }
+      bar: 'World',
+    },
   })
   @Injectable()
   class MyState {
@@ -35,8 +35,8 @@ describe('Selector', () => {
     name: 'zoo',
     defaults: {
       foo: 'Hello',
-      bar: 'World'
-    }
+      bar: 'World',
+    },
   })
   @Injectable()
   class MyState2 {
@@ -61,7 +61,7 @@ describe('Selector', () => {
   describe('(Decorator)', () => {
     it('should select the state', async(() => {
       TestBed.configureTestingModule({
-        imports: [NgxsModule.forRoot([MyState])]
+        imports: [NgxsModule.forRoot([MyState])],
       });
 
       const store: Store = TestBed.inject(Store);
@@ -71,7 +71,7 @@ describe('Selector', () => {
 
     it('should select using the meta selector', async(() => {
       TestBed.configureTestingModule({
-        imports: [NgxsModule.forRoot([MyState])]
+        imports: [NgxsModule.forRoot([MyState])],
       });
 
       const store: Store = TestBed.inject(Store);
@@ -81,7 +81,7 @@ describe('Selector', () => {
 
     it('should still be usable as a function', async(() => {
       TestBed.configureTestingModule({
-        imports: [NgxsModule.forRoot([MyState])]
+        imports: [NgxsModule.forRoot([MyState])],
       });
 
       const store: Store = TestBed.inject(Store);
@@ -92,7 +92,7 @@ describe('Selector', () => {
 
     it('should select multiples', async(() => {
       TestBed.configureTestingModule({
-        imports: [NgxsModule.forRoot([MyState, MyState2])]
+        imports: [NgxsModule.forRoot([MyState, MyState2])],
       });
 
       const store: Store = TestBed.inject(Store);
@@ -102,7 +102,7 @@ describe('Selector', () => {
 
     it('should select multiples from self and others', async(() => {
       TestBed.configureTestingModule({
-        imports: [NgxsModule.forRoot([MyState, MyState2])]
+        imports: [NgxsModule.forRoot([MyState, MyState2])],
       });
 
       const store: Store = TestBed.inject(Store);
@@ -114,8 +114,8 @@ describe('Selector', () => {
       @State<any>({
         name: 'counter',
         defaults: {
-          value: 0
-        }
+          value: 0,
+        },
       })
       @Injectable()
       class TestState {
@@ -133,7 +133,7 @@ describe('Selector', () => {
       }
 
       TestBed.configureTestingModule({
-        imports: [NgxsModule.forRoot([TestState])]
+        imports: [NgxsModule.forRoot([TestState])],
       });
 
       const store: Store = TestBed.inject(Store);
@@ -148,8 +148,8 @@ describe('Selector', () => {
           name: 'counter',
           defaults: {
             foo: 'Hello',
-            bar: 'World'
-          }
+            bar: 'World',
+          },
         })
         @Injectable()
         class TestState {
@@ -167,7 +167,7 @@ describe('Selector', () => {
         }
 
         TestBed.configureTestingModule({
-          imports: [NgxsModule.forRoot([TestState])]
+          imports: [NgxsModule.forRoot([TestState])],
         });
 
         const store: Store = TestBed.inject(Store);
@@ -186,8 +186,8 @@ describe('Selector', () => {
           name: 'counter',
           defaults: {
             foo: 'Hello',
-            bar: 'World'
-          }
+            bar: 'World',
+          },
         })
         @Injectable()
         class TestState {
@@ -202,7 +202,7 @@ describe('Selector', () => {
         }
 
         TestBed.configureTestingModule({
-          imports: [NgxsModule.forRoot([TestState])]
+          imports: [NgxsModule.forRoot([TestState])],
         });
 
         const store: Store = TestBed.inject(Store);
@@ -220,7 +220,7 @@ describe('Selector', () => {
   describe('(Selector Options)', () => {
     function setupStore(states: StateClass[], extendedOptions?: Partial<NgxsConfig>) {
       TestBed.configureTestingModule({
-        imports: [NgxsModule.forRoot(states, extendedOptions)]
+        imports: [NgxsModule.forRoot(states, extendedOptions)],
       });
       const store: Store = TestBed.inject(Store);
       return store;
@@ -231,8 +231,8 @@ describe('Selector', () => {
         name: 'zoo_1',
         defaults: {
           foo: 'Foo1',
-          bar: 'Bar1'
-        }
+          bar: 'Bar1',
+        },
       })
       @Injectable()
       // tslint:disable-next-line: class-name
@@ -257,8 +257,8 @@ describe('Selector', () => {
         name: 'zoo_2',
         defaults: {
           foo: 'Foo2',
-          bar: 'Bar2'
-        }
+          bar: 'Bar2',
+        },
       })
       @Injectable()
       // tslint:disable-next-line: class-name
@@ -288,8 +288,8 @@ describe('Selector', () => {
         // Arrange
         const store = setupStore([MyStateV4_1, MyStateV4_2], {
           selectorOptions: {
-            injectContainerState: false
-          }
+            injectContainerState: false,
+          },
         });
         // Act & Assert
         expect(store.selectSnapshot(MyStateV4_1.foo)).toBe('Foo1');
@@ -304,8 +304,8 @@ describe('Selector', () => {
         // Arrange
         const store = setupStore([MyStateV4_1, MyStateV4_2], {
           selectorOptions: {
-            suppressErrors: false
-          }
+            suppressErrors: false,
+          },
         });
         // Act
         let exception: Error | null = null;
@@ -324,12 +324,12 @@ describe('Selector', () => {
         name: 'zoo',
         defaults: {
           foo: 'Foo',
-          bar: 'Bar'
-        }
+          bar: 'Bar',
+        },
       })
       @SelectorOptions({
         injectContainerState: false,
-        suppressErrors: false
+        suppressErrors: false,
       })
       @Injectable()
       class MyStateV4 {
@@ -415,8 +415,8 @@ describe('Selector', () => {
         name: 'zoo',
         defaults: {
           foo: 'Foo',
-          bar: 'Bar'
-        }
+          bar: 'Bar',
+        },
       })
       @Injectable()
       class MyStateV4 {
@@ -433,7 +433,7 @@ describe('Selector', () => {
 
       @SelectorOptions({
         injectContainerState: false,
-        suppressErrors: false
+        suppressErrors: false,
       })
       class MyStateV4Queries {
         @Selector([MyStateV4, MyStateV4.foo])
@@ -490,8 +490,8 @@ describe('Selector', () => {
         name: 'zoo',
         defaults: {
           foo: 'Foo',
-          bar: 'Bar'
-        }
+          bar: 'Bar',
+        },
       })
       @Injectable()
       class MyStateV3 {
@@ -575,7 +575,7 @@ describe('Selector', () => {
   describe('(from createSelector)', () => {
     it('should select the state', async(() => {
       TestBed.configureTestingModule({
-        imports: [NgxsModule.forRoot([MyState])]
+        imports: [NgxsModule.forRoot([MyState])],
       });
 
       const store: Store = TestBed.inject(Store);
@@ -586,7 +586,7 @@ describe('Selector', () => {
 
     it('should allow for null in the returned value [regression fix]', async(() => {
       TestBed.configureTestingModule({
-        imports: [NgxsModule.forRoot([MyState])]
+        imports: [NgxsModule.forRoot([MyState])],
       });
 
       const store: Store = TestBed.inject(Store);
@@ -600,7 +600,7 @@ describe('Selector', () => {
 
     it('should allow for undefined in the returned value [regression fix]', async(() => {
       TestBed.configureTestingModule({
-        imports: [NgxsModule.forRoot([MyState])]
+        imports: [NgxsModule.forRoot([MyState])],
       });
 
       const store: Store = TestBed.inject(Store);
@@ -614,7 +614,7 @@ describe('Selector', () => {
 
     it('should select using the meta selector', async(() => {
       TestBed.configureTestingModule({
-        imports: [NgxsModule.forRoot([MyState])]
+        imports: [NgxsModule.forRoot([MyState])],
       });
 
       const store: Store = TestBed.inject(Store);
@@ -625,7 +625,7 @@ describe('Selector', () => {
 
     it('should still be usable as a function', async(() => {
       TestBed.configureTestingModule({
-        imports: [NgxsModule.forRoot([MyState])]
+        imports: [NgxsModule.forRoot([MyState])],
       });
 
       const store: Store = TestBed.inject(Store);
@@ -637,7 +637,7 @@ describe('Selector', () => {
 
     it('should select multiples', async(() => {
       TestBed.configureTestingModule({
-        imports: [NgxsModule.forRoot([MyState, MyState2])]
+        imports: [NgxsModule.forRoot([MyState, MyState2])],
       });
 
       const store: Store = TestBed.inject(Store);
@@ -657,14 +657,14 @@ describe('Selector', () => {
           name: 'counter',
           defaults: {
             foo: 'Hello',
-            bar: 'World'
-          }
+            bar: 'World',
+          },
         })
         @Injectable()
         class TestState {}
 
         TestBed.configureTestingModule({
-          imports: [NgxsModule.forRoot([TestState])]
+          imports: [NgxsModule.forRoot([TestState])],
         });
 
         const store: Store = TestBed.inject(Store);
@@ -692,14 +692,14 @@ describe('Selector', () => {
           name: 'counter',
           defaults: {
             foo: 'Hello',
-            bar: 'World'
-          }
+            bar: 'World',
+          },
         })
         @Injectable()
         class TestState {}
 
         TestBed.configureTestingModule({
-          imports: [NgxsModule.forRoot([TestState])]
+          imports: [NgxsModule.forRoot([TestState])],
         });
 
         const store: Store = TestBed.inject(Store);
@@ -724,7 +724,7 @@ describe('Selector', () => {
   describe('Errors in selectSnapshot', () => {
     @State<number[]>({
       name: 'tasks',
-      defaults: [1, 2, 3, 4]
+      defaults: [1, 2, 3, 4],
     })
     @Injectable()
     class TasksState {
@@ -740,7 +740,7 @@ describe('Selector', () => {
 
     it('should be a wrong mutation', () => {
       TestBed.configureTestingModule({
-        imports: [NgxsModule.forRoot([TasksState])]
+        imports: [NgxsModule.forRoot([TasksState])],
       });
 
       const store = TestBed.inject(Store);
@@ -754,7 +754,7 @@ describe('Selector', () => {
 
     it('should be incorrect mutation', () => {
       TestBed.configureTestingModule({
-        imports: [NgxsModule.forRoot([TasksState], { developmentMode: true })]
+        imports: [NgxsModule.forRoot([TasksState], { developmentMode: true })],
       });
 
       const store = TestBed.inject(Store);
@@ -771,9 +771,9 @@ describe('Selector', () => {
         imports: [
           NgxsModule.forRoot([TasksState], {
             developmentMode: true,
-            selectorOptions: { suppressErrors: false }
-          })
-        ]
+            selectorOptions: { suppressErrors: false },
+          }),
+        ],
       });
 
       const store = TestBed.inject(Store);
@@ -793,7 +793,7 @@ describe('Selector', () => {
   describe('Select errors', () => {
     @State<number[]>({
       name: 'tasks',
-      defaults: [1, 2, 3, 4]
+      defaults: [1, 2, 3, 4],
     })
     @Injectable()
     class NumberListState {
@@ -810,9 +810,9 @@ describe('Selector', () => {
         imports: [
           NgxsModule.forRoot([NumberListState], {
             developmentMode: true,
-            selectorOptions: { suppressErrors: false }
-          })
-        ]
+            selectorOptions: { suppressErrors: false },
+          }),
+        ],
       });
 
       const store = TestBed.inject(Store);
@@ -835,6 +835,9 @@ describe('Selector', () => {
     });
 
     it('@Selector should only work on methods', () => {
+      // Arrange
+      let message: string | null = null;
+      // Act
       try {
         class MyComponent {
           // @ts-ignore
@@ -842,9 +845,11 @@ describe('Selector', () => {
         }
 
         new MyComponent();
-      } catch ({ message }) {
-        expect(message).toEqual('Selectors only work on methods.');
+      } catch (error) {
+        message = error.message;
       }
+      // Assert
+      expect(message).toEqual('Object prototype may only be an Object or null: undefined');
     });
   });
 
@@ -866,14 +871,14 @@ describe('Selector', () => {
       name: 'contacts',
       defaults: {
         entities: {},
-        ids: []
-      }
+        ids: [],
+      },
     })
     @Injectable()
     class ContactsState {
       @Selector([ContactsState.ids, ContactsState.entityMap])
       static orderedContactNames(ids: number[], map: EntityMap<Contact>) {
-        return ids.map(id => map[id].name);
+        return ids.map((id) => map[id].name);
       }
 
       @Selector()
@@ -891,9 +896,9 @@ describe('Selector', () => {
       TestBed.configureTestingModule({
         imports: [
           NgxsModule.forRoot([ContactsState], {
-            selectorOptions: { suppressErrors: false, injectContainerState: false }
-          })
-        ]
+            selectorOptions: { suppressErrors: false, injectContainerState: false },
+          }),
+        ],
       });
 
       const store = TestBed.inject(Store);
@@ -910,10 +915,10 @@ describe('Selector', () => {
           entities: {
             456: { name: 'Artur' },
             234: { name: 'Mark' },
-            123: { name: 'Max' }
+            123: { name: 'Max' },
           },
-          ids: [234, 123, 456]
-        }
+          ids: [234, 123, 456],
+        },
       });
       // Act
       const result = store.selectSnapshot(ContactsState.orderedContactNames);